[FEATURE] Merged feat-schedule-pickup and feat-cancel-pickup

See merge request courses/2024-winter/csci5308/Group01!16

+++ src/main/java/com/group1/ecocredit/models/Waste.java
@@ -16,7 +16,7 @@ public class Waste {
     private Integer category_id;
 
     @Column(name = "pickup_id")
-    private int pickup_id;
+    private long pickup_id;
 
     private float weight;
